+++ heylogs-api/src/main/java/nbbrd/heylogs/Extractor.java
@@ -5,6 +5,7 @@ import com.vladsch.flexmark.ast.RefNode;
 import com.vladsch.flexmark.ast.Reference;
 import com.vladsch.flexmark.util.ast.Document;
 import com.vladsch.flexmark.util.ast.Node;
+import lombok.NonNull;
 
 import java.time.LocalDate;
 import java.time.Year;
@@ -14,10 +15,10 @@ import java.util.List;
 import java.util.regex.Pattern;
 
 @lombok.Value
-@lombok.Builder
-public class VersionFilter {
+@lombok.Builder(toBuilder = true)
+public class Extractor {
 
-    public static final VersionFilter DEFAULT = VersionFilter.builder().build();
+    public static final Extractor DEFAULT = Extractor.builder().build();
 
     @lombok.NonNull
     @lombok.Builder.Default
@@ -34,23 +35,26 @@ public class VersionFilter {
     @lombok.Builder.Default
     int limit = Integer.MAX_VALUE;
 
+    @lombok.Builder.Default
+    boolean ignoreContent = false;
+
     private boolean isUnreleasedPattern() {
         return unreleasedPattern.asPredicate().test(ref);
     }
 
-    private boolean containsRef(Version version) {
+    private boolean containsRef(@NonNull Version version) {
         return (isUnreleasedPattern() && version.isUnreleased()) || version.getRef().contains(ref);
     }
 
-    public boolean contains(Heading heading) {
-        return contains(Version.parse(heading));
-    }
-
-    public boolean contains(Version version) {
+    public boolean contains(@NonNull Version version) {
         return containsRef(version) && timeRange.contains(version.getDate());
     }
 
-    public void apply(Document root) {
+    public boolean contains(@NonNull Heading heading) {
+        return contains(Version.parse(heading));
+    }
+
+    public void extract(@NonNull Document root) {
         int found = 0;
         boolean keep = false;
 
@@ -59,7 +63,10 @@ public class VersionFilter {
 
         for (Node current : root.getChildren()) {
 
-            if (current instanceof Heading && Version.isVersionLevel((Heading) current)) {
+            boolean versionHeading = current instanceof Heading
+                    && Version.isVersionLevel((Heading) current);
+
+            if (versionHeading) {
                 if (found >= getLimit() || !contains((Heading) current)) {
                     keep = false;
                 } else {
@@ -73,6 +80,9 @@ public class VersionFilter {
                         .descendants(current)
                         .map(node -> node.getReference().toString())
                         .forEach(refNodes::add);
+                if (versionHeading && ignoreContent) {
+                    keep = false;
+                }
             } else {
                 if (current instanceof Reference) {
                     references.add((Reference) current);